feat: colonel-billing Zod schema + register stripe orgs response type
colonel-billing.ts: add ColonelBillingStripeOrgsResponse schema and
ColonelBillingStripeOrg record type (mirrors the new /colonel/billing/orgs
wire shape). Re-export from registry.

useAdminBilling: ESLint auto-fix + adapt to the new schema-driven type
for stripe org records.
